msm: ipc: Update the work item in the read path
IPC Message Router schedules a work item to read incoming messages
from the remote subsystems. Once a message is read, the work item
queues itself again in the workqueue. The queued work item performs
an uninterruptible wait for further messages. This causes the "loadavg"
to be consistently greater than 1. So refactor the read path such that
the read work item is queued only after the message has arrived.

CRs-Fixed: 383657
Change-Id: If532788de6377dd54574b57ec7a0779126d26e9b
Signed-off-by: Karthikeyan Ramasubramanian <kramasub@codeaurora.org>