msm: smd: Add receive interrupt masking support
SMD is used for communicating power state with the RPM processor.
During a power-down sequence, sleep information is transferred to the
RPM processor and the system blocks and polls for a response.  When the
RPM sends the response, it triggers an interrupt which then causes the
Application processor to abort the power-down sequence.

Add a new function that allows interrupts to be masked/unmasked for an
entire subsystem which allows the power code to mask interrupts while
powering down to prevent aborting the power-down sequence.

Change-Id: I52298f025032d47c7353dced6c1298766060ef0c
Signed-off-by: Eric Holmberg <eholmber@codeaurora.org>

Conflicts:
	arch/arm/mach-msm/include/mach/msm_smd.h