arm: common: fix cpaccess asm instruction
The write_val argument was not specified as an input
on the asm instruction. The compiler thought it wasn't used,
even though the asm instruction does use r0, which
is how write_val is passed into the function. The compiler
would optimise out the argument being passed into
the function.

Add write_val as an input operand to prevent it being
optimised out.

Change-Id: I71f35e1b2e3c655f4020163d089d9c71bbd6614b
Signed-off-by: Neil Leeder <nleeder@codeaurora.org>